[ACPI] ACPICA 20050309 from Bob Moore
The string-to-buffer implicit conversion code has been
modified again after a change to the ACPI specification.
In order to match the behavior of the other major ACPI
implementation, the target buffer is no longer truncated
if the source string is smaller than an existing target
buffer. This change requires an update to the ACPI spec,
and should eliminate the recent AE_AML_BUFFER_LIMIT issues.

The "implicit return" support was rewritten to a new
algorithm that solves the general case. Rather than
attempt to determine when a method is about to exit,
the result of every ASL operator is saved momentarily
until the very next ASL operator is executed. Therefore,
no matter how the method exits, there will always be a
saved implicit return value.  This feature is only enabled
with the acpi_gbl_enable_interpreter_slack flag which
Linux enables unless "acpi=strict".  This should
eliminate AE_AML_NO_RETURN_VALUE errors.

Implemented implicit conversion support for the predicate
(operand) of the If, Else, and While operators. String and
Buffer arguments are automatically converted to Integers.

Changed the string-to-integer conversion behavior to match
the new ACPI errata: "If no integer object exists, a new
integer is created. The ASCII string is interpreted as a
hexadecimal constant. Each string character is interpreted
as a hexadecimal value ('0'-'9', 'A'-'F', 'a', 'f'),
starting with the first character as the most significant
digit, and ending with the first non-hexadecimal character
or end-of-string."  This means that the first non-hex
character terminates the conversion and this is the code
that was changed.

Fixed a problem where the ObjectType operator would fail
(fault) when used on an Index of a Package which pointed
to a null package element.  The operator now properly
returns zero (Uninitialized) in this case.

Fixed a problem where the While operator used excessive
memory by not properly popping the result stack during
execution. There was no memory leak after execution,
however. (Code provided by Valery Podrezov.)

Fixed a problem where references to control methods within
Package objects caused the method to be invoked, instead
of producing a reference object pointing to the method.

Restructured and simplified the pswalk.c module
(acpi_ps_delete_parse_tree) to improve performance and reduce
code size. (Code provided by Alexey Starikovskiy.)

Signed-off-by: Len Brown <len.brown@intel.com>