[PATCH] USB: usbnet fix for Zaurus C-860
This patch resolves a recent problem with the Zaurus C-860 support.

A change to correct handling of Zaurii that are lying about their support
for the "CDC Ethernet" class specification broke the C-860, which tells
an entirely different lie (that it supports "CDC MDLM", providing access
to a cell phone modem).  The code expecting it to be telling a lie about
CDC Ethernet support naturally misbehaved.  (Sharp should straighten out
its story.  The 2.6 OpenZaurus kernels don't have any such issues...)

The fix is just to recognize this bogus MDLM stuff and ignore it.
This patch also includes the two MDLM descriptors in <linux/usb_cdc.h>
although they're not currently used.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>