[TG3]: Fix jumbo frames phy settings
Fix jumbo frame settings on all copper phys that support jumbo frames by
setting the fifo elasticity bit. This setting is for the phy's tx fifo to
properly handle jumbo frames. Note that a similar jumbo frame fix for the
phy's rx fifo was made to tg3 in the past.

Signed-off-by: Michael Chan <mchan@broadcom.com>
ACKed-by: Jeff Garzik <jgarzik@pobox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>