[TG3]: Fix ethtool set functions
Fix all relevant ethtool set functions to properly handle the
not-netif_running() case. In most cases, the new settings are accepted without
setting the hardware if not-netif_running(). The new settings will take effect
when the device is subsequently brought up. tg3_nway_reset() is the exception
where it will return -EAGAIN if not-netif_running().

Signed-off-by: Michael Chan <mchan@broadcom.com>
ACKed-by: Jeff Garzik <jgarzik@pobox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>