[PATCH] USB Storage: exit control thread immediately upon disconnect
This patch started life as as475 from Alan Stern.  It has been rediffed
against the tip, tho that was several days ago.

This patch causes the main control thread to exit as soon as possible,
i.e., as soon as the DISCONNECTING flag is set.  It no longer waits for an
explicit exit command, one with srb == NULL.

There won't be any bad implications for our interaction with the SCSI
midlayer, because once the DISCONNECTING flag is set we fail every
submitted command immediately in the queuecommand routine.  And if a
command manages to squeeze through the crack (submitted and accepted
before we disconnect but not yet processed), the SCSI midlayer will cancel
it automatically when we remove the host.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Matthew Dharm <mdharm-usb@one-eyed-alien.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>