[SERIAL] Set port.dev to PCMCIA device.
Since serial_cs provides power management callbacks, we should not use
the fallback methods in 8250.c.  We tell 8250.c about this when we
associate a struct device with the port, and, since we now have such
a beast, use it.

This also makes sysfs indicate which port is associated with which
PCMCIA device.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>