[SPARC64]: Support >=cheetah+ dual-dtlbs properly.
UltraSPARC-III+ (aka. cheetah+) and later chips have three
D-TLB units.  One is fully assosciative and has 16 entries,
the other two have 512 entries each and are 2 way subblocked
and then indexed by the PAGE_MASK bits.  This implies that
the two 512 entry TLBs need to know the page size in order
to index things properly.

The page size to use is encoded in page size fields within
the TLB context registers.

Up until this point we just left them at zero, which meant
that both 512 entry TLBs were used for 8K translations only.
The rest (including the 4MB kernel mappings) went into the
16 entry fully assosciative TLB.

Now, for the kernel, we use the first 512 entries for 4MB
mappings and the second 512 entries for PAGE_SIZE mappings
(ie. for vmalloc and modules).  For the user, we use the
first 512 entries for PAGE_SIZE and the second 512 entries
for HPAGE_SIZE if the user maps any hugetlb pages else we
use the second half for user PAGE_SIZE stuff as well.

Most of this changeset is clerical.  We move most of the
mm->context layout defines into asm/mmu.h, we abstract the
mm->context type so that it is now much easier to see and
trap accesses to the context values.  In particular, one
can find all context handling by grepping the sparc64 sources
for CTX_* and {PRIMARY,SECONDAY}_CONTEXT.

Happily, the fast paths of the kernel are mostly untouched
by this new stuff.  We add 2 cycles to trap entry, 1 cycle
to trap exit, and 2 cycles to the window trap fixup code.

Signed-off-by: David S. Miller <davem@davemloft.net>