[PATCH] Stallion driver module clean up
These two patches continue the work that Wayne Meissner started and are
against the current bk tree.

These patches allow the stallion driver to be built-in and loaded at boot
time, the current #ifdef MODULE only allows the init code to be included if
compiled as a module.

Tested for compile, boot and running on our console server as module and
built-in.

Signed-off-by Darren Williams <dsw@gelato.unsw.edu.au>

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>