[PATCH] checkstack: fix sort misbehavior for long function names
Fix sort behavior when long names are encountered.  The previour regular
expression depended on a tab to find the size in a string.  For long names,
this tab no longer exists, so it is smarter to check for a colon instead.

For the kernel, this change shouldn't make a difference.  But people
started using the same script for c++ code in other projects.

Signed-off-by: Jörn Engel <joern@wohnheim.fh-wedel.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>