[PATCH] Fix irq_affinity write from /proc for ia64
Made GENERIC_HARDIRQ mechanism work for ia64 and CPU hotplug.  When write
to /proc/irq is handled it is not appropriate to perform set_rte
immediatly, since there is a race when the interrupt is asserted while the
re-program is happening.  Hence such programming is only safe when we do
the re-program at the time of servicing an interrupt.  This got broken when
GENERIC_HARDIRQ got introduced for ia64.

- added CONFIG_PENDING_IRQ so default /proc/irq write handler can do the right
  thing.

TBD: We currently dont handle redirectable hint either in the display, or
when we handle writes to /proc/irq/XX/smp_affinity.  We need an arch
specific way to account for the presence of "r" hint when we handle the
proc write.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>
Cc: <linux-ia64@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>