[PATCH] Fix PCMCIA resume with card inserted
The following patch fixes a minor but annoying issue with PCMCIA resume.
When a machine is suspended with an empty socket, and then resumed with a
card in the socket, the card is ignored.  The following patch adjusts the
resume paths such that we detect the card and properly initialise it.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>