[PATCH] Add a scsi_device flag for RETRY_HWERROR
It turns out that a bunch of USB-IDE converters make the mistake of
returning SK = 04 (Hardware Error) whenever the IDE device signals any
sort of error, without bothering to distinguish recoverable from
non-recoverable errors.  The best way to handle this is for usb-storage to
set a per-device flag indicating that these errors should always be
retried.  The current scheme (blacklist flag but no per-device flag) isn't
well suited for this situation.

This patch adds the per-device flag and sets it initially based on the
blacklist setting.  Once this has been merged, a separate patch will be
submitted to Matt Dharm adding the corresponding support to usb-storage.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>