[PATCH] USB: add driver for CP2101/CP2102 RS232 adaptors
This patch against 2.6.10 provides the driver for the CP2101 and CP2102
USB to RS232 adaptor chips which are found in some serial converter
cables and usb mobile phone cables.

It provides cp2101.c, and patches Kconfig and Makefile in
drivers/usb/serial Since there is no technical data sheet freely
available for these chips, much of the code in this driver has been
based on protocol analysis, assumptions and guesswork.

The cp210x family of devices have OEM programmable device IDs so updates
will be required as and when new devices are found.

The driver currently only supports configuration of baud rate, data
bits, stop bits and parity mode. While this is enough to enable the use
of a wide range of devices based on this chip, the driver is still
incomplete in many areas such as flow control.  Unfortunately without
technical datasheets, implementing these features is made very difficult
indeed.

Signed-off-by: Craig Shelley craig@microtron.org.uk
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>