[PATCH] Kprobes: Allow/deny probes on int3/breakpoint instruction?
Kprobes did an improper exit when a probe is inserted on an int3
instruction.  In case of normal execution of int3/breakpoint instruction,
it oops!.  Probe on an int3 instruction was not handled properly by the
kprobes, it generated faults after oops!  doing an improper exit with
holding the lock.  This fix employes a bit different method to handle probe
on an int3/breakpoint instruction.

On execution of an int3/breakpoint instruction (placed by kprobe),
kprobes_handler() is called which sets it for single stepping in-line(it
does not matter whether we single step out-of-line/inline since the single
stepping instruction is same).  Now it single steps on int3/breakpoint
instruction here, entering kprobes_handler() once again.  Kprobes now
check's the status that it is single stepping and avoids the recursion.  It
runs down through the trap handler and oops messages is seen on the console
since it executed int3/breakpoint instruction.  Here the kprobes single
stepping handler never gets called.

Is this behaviour acceptable ?  Or should we avoid putting probes on an
int3 /breakpoint instruction ?  How should it handle such situations?
Below is the patch to allow probes on an int3/breakpoint instruction.

This patch fixes the above problem by doing a proper exit while avoiding
recursion.

Signed-off-by: Prasanna S Panchamukhi <prasanna@in.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>