[IA64] fix for unwind problem through dispatch_illegal_op_fault
Here is the patch worked out with David Mosberger to fix the unwind info
for dispatch_illegal_op_fault in ivt.S.  Before adding this patch, an
illegal op in the kernel would not show a complete stack trace in the
oops message.

Add PT_REGS_UNWIND_INFO to dispatch_illegal_op_fault to allow stack
unwind to continue through that stack level when kernel code commits an
illegal operation. 

Signed-off-by: Bob Montgomery <bob.montgomery@hp.com> 
Signed-off-by: Tony Luck <tony.luck@intel.com>